% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataconnector.R
\name{object2hdfs}
\alias{object2hdfs}
\title{Save an R object to a file in HDFS or the local file system.}
\usage{
object2hdfs(object, url, overwrite = FALSE,
  hdfsConfigurationFile = paste(system.file(package = "dataconnector"),
  "/conf/hdfs.json", sep = ""), hdfsConfigurationStr = "")
}
\arguments{
\item{object}{R object to save.}

\item{url}{File URL. Examples: '/tmp/file.txt', 'hdfs:///file.txt'.

                     We also support globbing. Examples: '/tmp/*.orc', 'hdfs:///tmp/*.orc'.}

\item{overwrite}{Overwrite file if it exists.}

\item{hdfsConfigurationFile}{By default: \code{paste(system.file(package='dataconnector'),'/conf/hdfs.json',sep='')}.

                             Options are:
                             \itemize{
                                 \item webhdfsPort: webhdfs port, integer
                                 \item hdfsPort: hdfs namenode port, integer
                                 \item hdfsHost: hdfs namenode host, string
                                 \item hdfsUser: hdfs username, string
                             }

                             An example file is:

                                 \{ \cr
                                 "webhdfsPort": 50070, \cr
                                 "hdfsPort": 9000, \cr
                                 "hdfsHost": "172.17.0.3", \cr
                                 "hdfsUser": "jorgem" \cr
                                 \}}

\item{hdfsConfigurationStr}{HDFS configuration in string format. Useful for Distributed R.
Normally users specify hdfsConfigurationFile instead of hdfsConfigurationStr.}
}
\value{
Nothing
}
\description{
Save an R object to a file in HDFS or the local file system.
}
\examples{
str <- "hello world"; object2hdfs(str, '/tmp/file.txt')
}

